[front] refactor: make the settings form generic + use more the store
