+++ frontend/src/pages/settings/preferences/Preferences.tsx
@@ -1,18 +1,48 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 
 import { Grid } from '@mui/material';
 
 import { ContentBox, ContentHeader, SettingsSection } from 'src/components';
-import UserSettingsForm from 'src/features/settings/preferences/videos/UserSettingsForm';
+import { useNotifications } from 'src/hooks';
+import GenericPollUserSettingsForm from 'src/features/settings/preferences/GenericPollUserSettingsForm';
 import SettingsMenu from 'src/features/settings/SettingsMenu';
+import { replaceSettings } from 'src/features/settings/userSettingsSlice';
 import {
   mainSectionBreakpoints,
   settingsMenuBreakpoints,
 } from 'src/pages/settings/layout';
+import { UsersService } from 'src/services/openapi';
 
-function PreferencePage() {
+const PreferencePage = () => {
   const { t } = useTranslation();
+  const dispatch = useDispatch();
+  const { contactAdministrator } = useNotifications();
+
+  /**
+   * In order to display the up-to-date user's preferences in all child
+   * components, we retrieve the user's settings from the API and refresh the
+   * Redux store.
+   */
+  useEffect(() => {
+    async function retrieveProfile() {
+      const response = await UsersService.usersMeSettingsRetrieve().catch(
+        () => {
+          contactAdministrator(
+            'error',
+            t('pollUserSettingsForm.errorOccurredWhileRetrievingPreferences')
+          );
+        }
+      );
+
+      if (response) {
+        dispatch(replaceSettings(response));
+      }
+    }
+
+    retrieveProfile();
+  }, [t, dispatch, contactAdministrator]);
 
   // subSectionBreakP can be changed independently of mainSectionBp
   const subSectionBreakpoints = mainSectionBreakpoints;
@@ -37,13 +67,13 @@ function PreferencePage() {
               title={t('preferences.preferencesOfThePoll')}
               {...subSectionBreakpoints}
             >
-              <UserSettingsForm />
+              <GenericPollUserSettingsForm pollName="videos" />
             </SettingsSection>
           </Grid>
         </Grid>
       </ContentBox>
     </>
   );
-}
+};
 
 export default PreferencePage;
